pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastmcp-server-collection"
version = "0.0.1"
description = "A collection of FastMCP servers for various APIs"
authors = [{ name = "Taylor Finklea", email = "taylor@novalark.ai" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "httpx==0.28.1",
  "mcp[cli]==1.9.0",
  "fastmcp==2.3.5",
  "python-dotenv==1.1.0",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Internet :: WWW/HTTP",
]
keywords = ["mcp", "fastmcp", "api", "servers", "skyvern"]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-cov"]

[project.urls]
"Homepage" = "https://github.com/TaylorFinklea/fastmcp-server-collection"
"Bug Tracker" = "https://github.com/TaylorFinklea/fastmcp-server-collection/issues"
"Documentation" = "https://github.com/TaylorFinklea/fastmcp-server-collection#readme"
